fradss/requirements.txt
flask==2.3.3
scikit-learn==1.3.0
rasterio==1.3.8
geopandas==0.13.2
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
folium==0.14.0
psycopg2-binary==2.9.9
